ADC and DAC logic rework & concerned SPI logic rework
The Top module have up to 8 ADC and DAC channels and can be configured.
The ADC and DAC ports are moved to the last 8 channels in the signal
bank.
Adapters use 16-bit ports and do padding/clipping internally.
SS pins of SPI system enumerate from 0 to 15 instead of 15 to 0.
Spi slaves are now specified by the FMC port they are at and their chip
index on the daughter board.
Part of SS distribution and MISO multiplexing are done in adapters.